prodigasistemas/gsan-batch#17 - Construcao de classes para geracao da parte 01

+++ src/main/java/br/gov/servicos/atendimentopublico/LigacaoEsgotoRepositorio.java
@@ -13,7 +13,7 @@ public class LigacaoEsgotoRepositorio {
 	@PersistenceContext
 	private EntityManager entity;
 	
-	public LigacaoEsgoto buscarLigacaoEsgotoPorIdImovel(Long idImovel){
+	public LigacaoEsgoto buscarLigacaoEsgotoPorIdImovel(Integer idImovel){
 		String sql = "select lig from LigacaoEsgoto lig where lig.imovel.id = :idImovel";
 		try {
 			return entity.createQuery(sql, LigacaoEsgoto.class)
