+++ src/main/java/br/gov/servicos/faturamento/DebitoCobrarRepositorio.java
@@ -34,7 +34,7 @@ public class DebitoCobrarRepositorio {
 		return debitos;
 	}
 	
-	public void reduzirParcelasCobradas(Integer referencia, Integer grupoFaturamento, List<Long> idsImoveis){
+	public void reduzirParcelasCobradas(Integer referencia, Integer grupoFaturamento, List<Integer> idsImoveis){
 		if (idsImoveis != null && idsImoveis.size() > 0){
 			StringBuilder sql = new StringBuilder();
 			sql.append(" update faturamento.debito_a_cobrar ")
