+++ src/test/java/br/gov/servicos/faturamento/CreditoRealizarRepositorioTestPreFaturamento.java
@@ -36,7 +36,7 @@ public class CreditoRealizarRepositorioTestPreFaturamento {
 	@ShouldMatchDataSet("credito_realizar_atualizacao_parcelas_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void atualizarParaImoveisDeContasSemRotaAlternativa() throws Exception{
-		List<Long> imoveis = contaRepositorio.imoveisDeContasSemRotaAlternativa(1, 201404, (short) 3, 1);
+		List<Integer> imoveis = contaRepositorio.imoveisDeContasSemRotaAlternativa(1, 201404, (short) 3, 1);
 		
 		repositorio.atualizarParcelas(201404, imoveis);
 		
@@ -47,7 +47,7 @@ public class CreditoRealizarRepositorioTestPreFaturamento {
 	@ShouldMatchDataSet("credito_realizar_atualizacao_residuo_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void atualizarParaImoveisDeContasComRotaAlternativa() throws Exception{
-		List<Long> imoveis = contaRepositorio.imoveisDeContasComRotaAlternativa(2, 201404, (short) 3, 1);
+		List<Integer> imoveis = contaRepositorio.imoveisDeContasComRotaAlternativa(2, 201404, (short) 3, 1);
 		
 		repositorio.atualizarValorResidual(imoveis);
 	}
