+++ src/test/java/br/gov/servicos/faturamento/DebitoCobrarRepositorioTest.java
@@ -46,7 +46,7 @@ public class DebitoCobrarRepositorioTest {
 	public void debitosCobrarVigentes() throws Exception {
 		logger.info("Verificar debitos a cobrar vigentes");
 		Imovel imovel = new Imovel();
-		imovel.setId(1L);
+		imovel.setId(1);
 
 		Collection<DebitoCobrar> debitos = repositorio.debitosCobrarPorImovelComPendenciaESemRevisao(imovel);
 		
@@ -60,7 +60,7 @@ public class DebitoCobrarRepositorioTest {
 	@ShouldMatchDataSet("debito_cobrar_atualizacao_contas_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void atualizarParaImoveisDeContasSemRotaAlternativa() throws Exception{
-		List<Long> imoveis = contaRepositorio.imoveisDeContasSemRotaAlternativa(1, 201404, (short) 3, 1);
+		List<Integer> imoveis = contaRepositorio.imoveisDeContasSemRotaAlternativa(1, 201404, (short) 3, 1);
 		
 		repositorio.reduzirParcelasCobradas(201404, 1, imoveis);
 	}
@@ -70,7 +70,7 @@ public class DebitoCobrarRepositorioTest {
 	@ShouldMatchDataSet("debito_cobrar_atualizacao_contas_alternativas_expected.yml")
 	@Cleanup(phase = TestExecutionPhase.AFTER, strategy = CleanupStrategy.USED_ROWS_ONLY)
 	public void atualizarParaImoveisDeContasComRotaAlternativa() throws Exception{
-		List<Long> imoveis = contaRepositorio.imoveisDeContasComRotaAlternativa(2, 201404, (short) 3, 1);
+		List<Integer> imoveis = contaRepositorio.imoveisDeContasComRotaAlternativa(2, 201404, (short) 3, 1);
 		
 		repositorio.reduzirParcelasCobradas(201404, 1, imoveis);
